[go] Consolidate example and benchmark code
